refactor: FileControls into 4 files
